ipacm: Fix to handled incorrect Mux Id mapping issue
Due to VLAN-PDN muxup sequence change like non-vlan
vlan4092 then vlan2 and mux_up up non-vlan, vlan2
then vlan4092, due to which Actual vlan2 Mux_Id muxed
with vlan4092 Mux_Id.

+++ ipacm/inc/IPACM_Xml.h
@@ -412,7 +412,7 @@ typedef struct
 typedef struct
 {
 	bool privateIPForwarding_enable;
-	uint8_t vlan; //0, if APN 1 or untagged traffic has to be accelerated, and the vlan ID, if tagged traffic, and that vlan has to be accelerated
+	uint16_t vlan; //0, if APN 1 or untagged traffic has to be accelerated, and the vlan ID, if tagged traffic, and that vlan has to be accelerated
 	char interface_name[IPA_IFACE_NAME_LEN];
 	char bridge_name[IPA_IFACE_NAME_LEN];
 	uint32_t bridge_net_mask;
